The class EmulateSubnet that ties together ospf instances would directly from the "send" method of one instance call the "receive" method on another instance. Now the "sends" are queued and read "recived" from the queue in a different context.  Thus ensuring that OSPF instances are not in each others call graphs, which can cause re-entrancy problems.
Previously when the verbose flag was used only sent packets were
traced. Now if the level is greater than 1 (-l 1) the recieved packets
are also traced.

Another flag "-s" has been added that can be used to supress one side
of the conversation. For example (-s ospf1 or -s ospf2).


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6750 26435e26-4b25-483a-942d-477236992e84