Bug fix in calculating the deleted part of the configuration tree when loading the configuration from a file.
E.g, if the current configuration contains:

rtrmgr {
    ...
}

and if we load configuration that contains the same rtrmgr configuration
block plus an "interfaces" configuration block:

interfaces {
    ...
}

rtrmgr {
    ...
}

The the load command fails with an error message like:
Node "rtrmgr" has no configured mandatory child node "targetname"

The root of the problem is that inside method
ConfigTreeNode::retain_different_nodes()
the "rtrmgr" nodes are considered different because they have
different node number in both configurations.

The solution is to introduce and use a new method
ConfigTreeNode::is_same() that can ignore the nodenum in the
comparison.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6806 26435e26-4b25-483a-942d-477236992e84