A clique of six routers with a seventh router feeding in a single route and then withdrawing the route was causing BGP to fail. The failure was either the fatal message "Unknown nexthop" from the next hop resolver or the next hop routing would dereference an iterator that was pointing at end.
The first problem is that on a slow system with an unresponsive RIB
many registrations can be queued. Test9 in test next hop resolver
demonstrates the problem. While waiting for the RIB, requests were
queued keyed on the subnet associated with the next hop that was being
requested. Two registrations followed by two deregistrations (with the
same subnet) were failing as all the state was being removed by the
first deregistration. The state is now stored in a multiset and the
regression test and the six router test seems to work.

The next hop resolver has also had some debugging added.

The route table nh lookup has methods to track the two queues which it
holds so that both queues can be handled in a consistent manor.

Bug found by: Dario Vieira


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6811 26435e26-4b25-483a-942d-477236992e84