The decoding of a link state update packet contaning multiple LSAs was failing.
The LSA decoder routines are passed all of the buffer. They are
supposd to return to the caller the amount of the buffer that has been
consumed. They were not adjusting the length of the buffer. This
problem has now been corrected.

The test program now has a link state update packet with multiple
LSAs.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6840 26435e26-4b25-483a-942d-477236992e84