The DR election had become broken. In a neighbour get_candidate_id() returns the neighbours router ID or interface ID as a function of protocol and linktype. In a peer get_candidate_id() returns the routers own ID. In event_hello_recived() _peer.get_candidate_id() had been incorrectly switched to get_candidate_id().
In event_hello_received() the DR and BDR would be recomputed when the
first hello packet was recieved from a neighbour instead of the first
time the state goes to TwoWay.

Previously event_wait_timer() called event_backup_seen() copied the
code from event_backup_seen() into event_wait_timer() so as not to be
confused by the debugging output.

When retransmitting an LSA don't update the age if its already at
MaxAge.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6857 26435e26-4b25-483a-942d-477236992e84