First part of the routing computation. The incremental computation is conditionally compiled out. All the relevant routing methods now take the SPT as an argument rather than operating on the SPT class variable.
An LSA being rejected on input because it has arrived less than a
second (MinLSArrival) after the previous LSA now generates a trace
message if input error tracing is enabled.

If a self originating LSA is requested from the database by the
arrival of a link state request packet the the delay queue is
fired. This should stop the problem of a Router-LSA being sent
immediately to a neighbour as soon as the adjacency becomes full.

Some additional debugging.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6953 26435e26-4b25-483a-942d-477236992e84