On FreeBSD 4.x, we can't check for immediate loopback connection, as socket connects/accepts get handled at splnet() in the kernel, so we have to give some slack to select() -- the accept/connect won't get serviced until we enter a wait state, and the kernel isn't preemptible.
On Linux and FreeBSD 5.x+, the connects are handled by netisrs,
and the kernel *may* preempt the process.

So don't try to use comm_sock_is_connected() for immediate connection
notification on anything which isn't Linux or FreeBSD 5.x
for the time being.

This is a very ugly bandaid for a very ugly problem -- but it
fixes the regression in test_peering1 test30 on but.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6979 26435e26-4b25-483a-942d-477236992e84