Its encode() not update().
If a self originated LSA arrives which has a higher sequence number
than our current database copy; then it is necessary to copy this
sequence number plus one into our LSA and send it back.

We did this correctly but update() had not been called to write the
sequence number back into the packet.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6997 26435e26-4b25-483a-942d-477236992e84