The IO subsystem now has to provide three methods, startup, running and shutdown.
Move the calls to the startup and shutdown methods out of the
constuctor for XrlIO. They are now called by the main class.

On shutdown the process now returns a status of shutdown and waits for
the IO subsystem to shutdown.

Routing tables are now created and deleted from the RIB.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7011 26435e26-4b25-483a-942d-477236992e84