When an interface is taken down all the neighbours are sent a kill neighbour event. If the router is the designated router for this interface then the neighbour makes an upcall to get itself removed from neighbour list in the Network-LSA.
Previously the determination of whether this router was the designated
router was made by checking the hello packet. Independent of the state
of the interface the neighbour could consider itself to be the
designated router. An attempt was being to remove the Network-LSA
twice. Which was causing the tinderbox failure and making it
impossible to take down an interface that was the designated router.

The methods for checking whether the interface is the designated
router or the backup designated router have been moved to the Peer
class and they now check the state of the interface not the contents
of the hello packet.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7021 26435e26-4b25-483a-942d-477236992e84