Start the hello timer on entering the "Waiting" state.
It is now fatal to receive the "WaitTimer" event in any state other
than "Waiting".

It is now fatal to receive the "BackupSeen" event in any state other
than "Waiting".

Event "BackupSeen" now stops the wait timer.

The designated router computation is run when there is a "BackupSeen"
or "NeighbourChange" event.

The "BackupSeen" event is only valid in state "Waiting", the
"NeighbourChange" event is only valid in states "Backup", "DR" or
"DRother".

Previously the event "BackupSeen" would never have been generated. It
was also possible for a "NeighbourChange" event to happen in state
"Waiting".


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7045 26435e26-4b25-483a-942d-477236992e84