When bringing up an interface set running to true as the first operation, otherwise it appears that the interface is down.
When adding a neighbour in the PointToPoint case only allow a single
neighbour.

Assert that the router does not believe it is the designated router
and backup designated router at the same time.

Start the hello timer on a PointToPoint interface.

Don't compute the router links if there are no neighbours.

When calling the area router with incoming LSAs use is_DR() and is_BDR().

Complete the method that get the neighbours address on a PointToPoint link.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7057 26435e26-4b25-483a-942d-477236992e84