Deal correctly with connection collisions.
Previously if an outbound connection was being attempted when an
incoming connection was received the incoming connection was
unconditionally dropped.

The problem could be reproduced by starting two BGP processes in IMUNES
with a 500ms delay between them. The processes would start together
and both connections would be dropped. Two minutes later the same
problem would recur...

The protocol requires that the IDs within the open message are checked
and the connection initiated from the peer with the higher ID should
be kept.

A separate class has been added that allows two TCP session peer
peering while the connection collision issue is being
resolved. Basically a second partial state machine.

All the timers in BGP are now stored in seconds not milliseconds.

The get_message routine now gets a pointer to the SocketClient
allowing an extra sanity check.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7090 26435e26-4b25-483a-942d-477236992e84