The original RouteEntry is provided even when withdrawing a summary to allow the peer_manager to detect if the route was a candidate for summarisation.
The routes are also going to be stored in the peer manager. Thus when
new areas are introduced the summarisation information can be fed into
the new area. Also if a stub area that was previously not performing
summarisation changes state all the route can be fed to the area on
demand.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7142 26435e26-4b25-483a-942d-477236992e84