Refactor the initialisation of links that are included in Router-LSAs.
In the case of Point-2-Point links set the link id to the neighbours
router ID and not its interface address. Also generate a stub link
that describes the network or host route to the neighbour, the stub
link was previously being generated but not making it to the
Router-LSA.

Comment out get_p2p_neighbour_address() which is no longer required.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7212 26435e26-4b25-483a-942d-477236992e84