Send all AS-External-LSAs that are received from the wire upto the peer manager to be sent to other areas.
No longer deal with an AS-External-LSA reaching MaxAge, previously the
timer management for AS-External-LSAs would be performed by areas
themselves. It is now going to managed externally to deal with the
case that an AS-External-LSA arrives and is propogated to other areas
and then the area is removed, previously the timer routines would have
pointed at the deleted area.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7264 26435e26-4b25-483a-942d-477236992e84