A new method "refresh_router_lsas" that causes a new Router-LSA to be generated in all areas.
The method "refresh_router_lsa" is now externally visable allowing
external entities to force the generation of a new Router-LSA.

A Router-LSA carries three flags; is this a border router, is this an
AS boudary router, is this a virtual link endpoint.

If the states of any of these flags changes a new Router-LSA has to be
generated.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7310 26435e26-4b25-483a-942d-477236992e84