support for a map protocol<->process.
i.e. the protocol in policy no longer needs to be the XORP process name.  Interanlly,
the policy uses only the logical name.  The XORP process name is used only in XRL stuff.

static_routes are now referred to as "static" in policy for example.

To use this feature execute this as the FIRST command in policy [before varmap]:
%create: xrl "$(policy.targetname)/policy/0.1/set_proto_target?protocol:txt=static&target:txt=static_routes";

it should be quite self explanatory


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7334 26435e26-4b25-483a-942d-477236992e84