A Router-LSA for the router itself always exists unconditionally in the area database. An LSA in the database can be marked as unavailable, an example would be no router links.
If an area is configured and all the interfaces are disabled then the
Router-LSa is not available. The get_lsa() interface used by the show
database operational command was not checking for availability and
returning all LSAs unconditionally. The LSA code itself checks for
availability before returning an LSA and the assert in this code was
being triggered.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7374 26435e26-4b25-483a-942d-477236992e84