When the router was the designated router and a neighbouring router on a broadcast interface was a border router or an AS boundary router a host route to the router itself was being installed.
R1 <--> N1 <--> R2

Router R1 generates a Network-LSA (N1) that is the next vertex towards
R2. the Vertex N1 has as its nexthop R1. A route to R2 was therefore
returning R1 as the nexthop.

When the router is a designated router on an interface put in direct
edges to each neighbour.

Allow host routes for router vertexs.

When processing AS-External-LSAs use the advertising router field to
perform the lookup in the routing table not the network. The
forwarding entry in the LSA is now being honoured.

Add a regression test routing4() that verifies that the correct routes
are introduced.

Bugzilla URL:	http://www.xorp.org/bugzilla/show_bug.cgi?id=295
Bug found by:	Patrick Marc Preuss


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7402 26435e26-4b25-483a-942d-477236992e84