Refactor the mechanism for sending the commands to the remote targets by introducing a centralized queue for all commands.
We need this centralized mechanism to ensure that the commands
are sent to the remote targets in the order the targets are
listed. Otherwise, there could be a race condition if some of
the targets try to communicate with each other immediately after
they receive the updates.

For example, if an IP address is moved from one interface to
another, the static_routes target may receive the interface update
and send a route update to RIB before the RIB has even received the
interface update. The result of such race condition is unpredictable:
e.g., routes with incorrect next-hop interface, or routes that
are rejected by the RIB.

This fixes the remaining issue in bugzilla entry #297

Bugzilla URL:	http://www.xorp.org/bugzilla/show_bug.cgi?id=297


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7437 26435e26-4b25-483a-942d-477236992e84