Add policy-based support to remove MEDs from the routes advertised by BGP to its peers.
The new BGP specific policy term is in the "then" block is the boolean
med-remove which if set to true will remove the MEDs from the
UPDATEs. E.g., to remove the MEDs from all UPDATEs:

protocols {
    bgp {
        ...
        export: "bgp-meds"
    }
}

policy {
    policy-statement bgp-meds {
        term foo1 {
            then {
                /* med: 10 */
                med-remove: true
            }
        }
    }
}

Note that "med-remove" has priority over the "med" modifier.
I.e., if "med: 10" is uncommented, then the MEDs are still removed.
Also, note that similar to the rest of the BGP policy code, if the
med-remove's value is changed, all UPDATEs are resent with the MEDs
being added/deleted as appropriate.

Requested by:	Atanu Ghosh


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7505 26435e26-4b25-483a-942d-477236992e84