Teach comm_sock_pair() to support AF_INET6 on Windows systems.
It is a peculiarity of Windows networking that at least one physical
interface on the system must be configured with the desired address
family in order for a loopback interface to be present (AF_INET, AF_INET6).

Note that if AF_UNIX is requested, AF_INET is still used.

Use sockaddr_storage. Update comments.

Obtained from:	XORP/Win32


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7506 26435e26-4b25-483a-942d-477236992e84