If the state transition to ExStart is normal then send a data description packet immediately and start the retransmit timer.
If the state transition to ExStart is due to an error just start the
retransmit timer. Previously an error in the database exchange could
cause back to back database description packets to be sent.

Its possible this was the cause of the bug below.

Bugzilla URL:	http://www.xorp.org/bugzilla/show_bug.cgi?id=292


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7564 26435e26-4b25-483a-942d-477236992e84