Add static filters for well known communities: NO_EXPORT, NO_ADVERTISE. Partial support for NO_EXPORT_SUBCONFED. Without these we wouldn't comply with RFC 1997.
Fix a bug where we would abort when we got an AS_PATH containing a
AS_CONFED_SET or AS_CONFED_SEQUENCE (which aren't yet supported), or
any other undefined AS_SEGMENT type.  Now we do the right thing, which
is to send a Notification, and drop the peering.

None of this has test suites yet, but it does pass all the tinderbox
tests.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7580 26435e26-4b25-483a-942d-477236992e84