When originating an aggregate route set the filter generation id of the aggregate to the genid value obtained from the first component route that triggered the generation of this aggregate.  Required for proper operation of export filters.
This can be safe if and only if all the existing routes are deleted prior
to installation of new version of the policy filters.  In other words, my
assumption is that it shouldn't ever be possible to see two component
routes in the aggregation table with two different filter genids at the
same time.  Correct / wrong?


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7639 26435e26-4b25-483a-942d-477236992e84