Allow route reflectors and confederations to be configured. It was possible to configure a route reflector client previously but once a peer was configured as a route reflector the state could not be changed.
The constructor for a bgp peer data no longer takes the type of the
peering. The determination of the type of peer computed by calling
compute_peer_type(). The local data structure is now bound to the peer
data structure to aid in the computation of the peer type. The AS
number of the router is now found by calling my_AS_number() in the
peer data structure, if a router is configured as part of a
confederation it has two AS numbers one that is used with other
confederation members and one that is used with other E-BGP peers,
hence the requirement to get the routers own AS number from the peer
data structure.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7696 26435e26-4b25-483a-942d-477236992e84