Implement Optional Session Attributes: AllowAutomaticStart, DampPeerOscillations, IdleHoldTime, IdleHoldTimer
draft-ietf-idr-bgp4-26.txt:
----------------------------------------
An example of DampPeerOscillations logic is an increase of the
IdleHoldTime value if a BGP peer oscillates connectivity
(connected/disconnected) repeatedly within a time period. To engage
this logic, a peer could connect and disconnect 10 times within 5
minutes. The IdleHoldTime value would be reset from 0 to 120 seconds.
----------------------------------------

The RFC leaves it up to the implementation to decide how to damp peer
oscillations, but offers the above as a possible implementation.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7725 26435e26-4b25-483a-942d-477236992e84