Check that the 16 byte authentication marker at the beginning of every bgp message is set to ones and if is isn't generate a notification message (Connection Not Synchronized). There is never any useful data here total waste of CPU to check this but the specification requires it.
Previously when an error occurred while parsing a message a
notification message would be scheduled to be sent to the
peer. Unfortunately the very next action was to flush the output
queue. The code is now back to waiting 10 seconds in state stopped
while the notification message is sent. If the connection no longer
exists then the state goes immediately to idle.

There was probably a bug in the catch clause it wasn't using a
reference.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7747 26435e26-4b25-483a-942d-477236992e84