Second attempt to fix test10.
BGP used to allow an identifier of 0.0.0.0 but this is not a legal
unicast IPv4 address. Recently BGP was fixed to return a error when
confronted with such a identifier. Modify the test to expect the error
not a successful session.

Remove all the local definitions of the notification error codes and
use the definitions in notification_codes.sh.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7785 26435e26-4b25-483a-942d-477236992e84