An AS_PATH is made up of segments either AS_SEQUENCE or AS_SET.
A string of this form used to a number of seqments of type AS_SEQUENCE.

     "1,2,(3,4,5),6,(7,8),9"

When the confederation support was added the meaning of '(' changed
from begining of AS_SEQUENCE to begining of AS_CONFED_SEQUENCE.

Recently code was added to verify that received update messages do not
contain AS_CONFED_SEQUENCE or AS_CONFED_SET in the received AS_PATH if
the router is not configured to be in a confederation. A notification
message is send and the connection torn down if this error
occurs. Hence a large number of the regression tests have been
failing.

Now AS_SEQUENCE is started with a '['.

     "1,2,[3,4,5],6,[7,8],9"


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7850 26435e26-4b25-483a-942d-477236992e84