The V bit in a Router-LSA should be set if any virtual link in the transit area is fully adjacent.
The V bit is the Router-LSA was being correctly set but it would only
be transmitted if the LSA was changed (router link change) or
LSRefreshTime was reached.

When any neighbour becomes fully adjacent or an adjacency is lost the
peer manager is notified. If the adjacency is associated with a
virtual link and the count of fully adjacent virtual links for this
transit area transitions between 0 and 1, the area router is prompted
to generate a new Router-LSA.

Router-LSAs should now be generated with the V bit set or unset
immediately the adjacency transition occurs.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7947 26435e26-4b25-483a-942d-477236992e84