When a route is deleted the associated Summary-LSAs that are sent to other areas are withdrawn. It is possible to compute if a Summary-LSA was prevously being generated and hence only withdraw if necessary.
Previously even if a Summary-LSA should not have been announced a
check was being made to see if the Summary-LSA was in the database and
its presence was considered a fatal error.

If an area range is manually reconfigured then its possible that the
computed state is incorrect. Print a warning message if an
inconsistency is discovered, should only be seen after a manual
reconfigure at any other time it may be flagging a problem.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7975 26435e26-4b25-483a-942d-477236992e84