If the router generates a new Router-LSA or Network-LSA then the topology has changed and routing table may need to be updated, unless the LSA is being generated due to a timer event.
In the case of a Router-LSA a change was already scheduling a routing
computation but now a timer event won't schedule a routing comptation.

A Network-LSA was not scheduling a routing computation, it is now and
a timer event will be ignored.

Bugzilla URL:	http://www.xorp.org/bugzilla/show_bug.cgi?id=468


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8032 26435e26-4b25-483a-942d-477236992e84