Routes provided by the policy framework (typically redistribution) or generated by the network statement are considered to be self originated.
Routes sent to an E-BGP peer have their nexthop rewritten to match the
local ip address of the peering. The exception was if the route was
self originated. This restriction has been removed so all nexthops on
E-BGP peers are now unconditionally rewritten.

The ability to rerwrite nexthops via policy has temporarily been
removed from the template as it would be ignored (bug 500).

Currently no check is made to exclude the rewritting of nexthops that
fall in the same subnet as the nexthop.

Bugzilla URL:	http://www.xorp.org/bugzilla/show_bug.cgi?id=472


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8046 26435e26-4b25-483a-942d-477236992e84