Add front-end support for configuring and managing the RIP MD5 keys. Now we can configure up to 256 keys with different password each, and start and end time:
authentication {
    type: "md5"
    md5 1 {		/* KeyID: [0, 255] */
        password: "FOO"
        start-time: "YYYY-MM-DD.HH:MM"
        end-time: "YYYY-MM-DD.HH:MM"
    }
}

Both start-time and end-time are optional. If start-time is empty,
then the key is activated immediately.
If the end time is empty or is same as the start time, then the
key will never expire.

Note that for the time being we still support the original md5
configuration, but this backward compatibility may be removed
to reduce confusion:

authentication {
    type: "md5"
    password: "FOO"
}

As a side-effect of this backward compatibility, the following
configuration will be a silent no-op:

authentication {
    type: "md5"
    password: ""
}

It appears that the new front-end is working, but more testings
are needed.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8060 26435e26-4b25-483a-942d-477236992e84