Impose an XRL input buffer size limit of 64KiB on what is advertised.
Previously the Finder server code would trust the value sent to
it by the remote XRL peer upon connection. This meant that the
Finder could easily be crashed by connecting to it with telnet
and sending random ASCII keystrokes.

These keystrokes would be interpreted as a 32-bit integer in network
byte order and then used to size the input buffer, causing a
memory allocation, creating the possibility of a std::bad_alloc
exception being thrown and not caught.

This fix is intended as a bandaid for a wider set of security
issues within XRL.

Bugzilla URL:	http://www.xorp.org/bugzilla/show_bug.cgi?id=143
Reviewed by:	atanu, pavlin


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8096 26435e26-4b25-483a-942d-477236992e84