Fix the MD5 key management backend.
Now the received packets' seqno information (and the flag whether
a packet has been received) are kept per key per (neighbor's) IPv4
address. The reason for the change is because the packets' seqno space
is per neighbor, but the keys are per "port" (i.e., subnet), so
previously the mechanism wouldn't work if there was more than one
MD5-enabled RIP router on the same subnet.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8112 26435e26-4b25-483a-942d-477236992e84