Refactor the authentication mechanism (both the front-end and the back-end) so now we support MD5 keys management and we are more consistent with the RIP implementation and the RIP front-end configuration.
Now we can configure up to 256 MD5 keys with different password
each, and start and end time:

authentication {
    md5 1 {             /* KeyID: [0, 255] */
        password: "FOO"
        start-time: "YYYY-MM-DD.HH:MM"
        end-time: "YYYY-MM-DD.HH:MM"
    }
}

Both start-time and end-time are optional. If start-time is empty,
then the key is activated immediately.
If the end time is empty or is same as the start time, then the
key will never expire.

In addition, the simple password configuration is:

authentication {
    simple-password: "FOO"
}

Note that the MD5 and simple password authentications are mutually
exclusive so they shouldn't be mixed. Unfortunately, for the time
being we don't have rtrmgr/xorpsh support that enforces the rule
that the user doesn't configure both of them.

Also, note that in the back-end implementation the textual name
of the plain-text/simple-password authentication handler has
been changed from "plaintext" to "simple" (for consistency with
the RIP implementation and with other vendors' implementations).
From user's perspective it doesn't matter (anymore), but the developers
should be aware of the change.

Just for the record, the old front-end for setting the authentication
was:

authentication {
    type: "plaintext"
    password: "FOO"
}

and

authentication {
    type: "md5"
    password: "FOO"
}

The old front-end is %deprecated in the rtrmgr template file.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8120 26435e26-4b25-483a-942d-477236992e84