BGP requires that update messages sent to I-BGP peers should have the local preference attribute.
When an update message arrives from a peer if it didn't come from an
IBGP peer it has a local preference added in the inbound filter
bank. In the outbound filter bank if the peering is not an IBGP
peering the local preference is removed.

Routes announced (redistribution) by BGP that are sent to an I-BGP
peering did not have a local preference attribute as the local
preference was not being added on input.

Annouced routes now have a local preference attribute added with the
default value.

The tests "test56" and "test56_ipv6" verify that the local preference
is being correctly added.

May also fix: http://www.xorp.org/bugzilla/show_bug.cgi?id=416

Bugzilla URL:	http://www.xorp.org/bugzilla/show_bug.cgi?id=517


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8124 26435e26-4b25-483a-942d-477236992e84