RFC 2328 Section 9.4 Electing the Designated Router (3): Requires that if no router has declared itself the designated router that the newly elected backup designated router should be used.
If no router had declared itself the designated router the router was
incorrectly electing itself.

If the router was started and a DR and BDR already existed if the
hello packet from the DR was seen first followed by the BDR a
backupseen event would be generated and the election would work
correctly.

If the router was started and a DR and BDR already existed if the
hello packet from the BDR was seen first a backupseen event would be
generated and the election would incorrectly elect itself as the DR.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8183 26435e26-4b25-483a-942d-477236992e84