In OSPFv2 the DR and BDR in the hello packet are interface addresses. In OSPFv3 the DR and BDR are router IDs.
The designated router election uses the router ID to resolve conflicts
if the priorities of two routers are the same. Unfortunately in the
OSPFv2 case the interface addresses were being compared not the router
IDs.

This problem was hidden as typically one the interface addresses is
choosen as the router ID.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8184 26435e26-4b25-483a-942d-477236992e84