When route flap damping is enabled a trie in the route flap damping table holds damping state. If damping is not enabled then there is no requirement to hold any state.
If damping is enabled after routes have flowed through the system then
when a replace or delete route arrives the orignal route may not be
present in the damping trie. Previously the code assumed that the
previous route must be in the system. Now if a route is not found in
the trie it is processed normally.

Two test programs test2 and test3 in test_route_flap_damping1.sh
demonstrate the problem.

Bugzilla URL:	http://www.xorp.org/bugzilla/show_bug.cgi?id=471


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8226 26435e26-4b25-483a-942d-477236992e84