If a link type was configured as point-to-point ("p2p") and no neighbour was configured then assert was being hit. The incorrect assumption was that if point-to-point was configured then a neighbour was also configured.
The simplest fix before the 1.2 release is to change the assert to an
error message. A neighbour must still be configured for point-to-point
to work. Also adding/removing a neighbour will now update the
Router-LSA.

A better fix would be derive the neighbour information as suggested by
Hasso Tepper.

Bugzilla URL:	http://www.xorp.org/bugzilla/show_bug.cgi?id=554


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8229 26435e26-4b25-483a-942d-477236992e84