The previous change was broken, and could result in the Dump failing to complete of the connection to the peer ever got backlogged (which is normal).
This fix reverses the sense of the test, to avoid dumping on triggered
events.  This is the same fix as before, but using a more reliable
test (we know when the event is triggered, but we don't always know
for sure when it's on a timer).

Note to anyone listening: test_peering2.sh uses test data from
../../../data, which is a separate CVS module.  If it doesn't exist,
the test still passes, but some of the code has not been tested, which
is how I missed the error this corrects.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8370 26435e26-4b25-483a-942d-477236992e84