Resolve a critical problem with multicast membership on raw sockets for the Microsoft Windows Server port of XORP 1.2.
On Windows, according to the Winsock documentation, it is required
that a raw socket be bound to an interface before it may be joined
to a multicast group. This however is problematic because the FEA
raw socket wrapper assumes that a single raw socket may be used to
transmit and/or receive on multiple interfaces (this is the UNIX
behaviour).

After consultation with Microsoft it appears that a bind() to
INADDR_ANY may be used to allow the IGMP joins via setsockopt() to
work. This was verified with Ethereal in a VMware virtual machine
running XORP and using OpenVPN/VMware interfaces.

Until the code is taught to ignore the error, a log message of the
following form may be seen when the OSPF protocol is used under Windows:
%%%
[ 2006/03/14 05:49:13 WARNING c:\xorp-1.2\fea\xorp_fea.exe FEA ] bind() failed:
An invalid argument was supplied.
%%%

Approved by:	atanu (implicitly)
Submitted by:	pavlin
Bugzilla ID:	http://www.xorp.org/bugzilla/show_bug.cgi?id=534


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8371 26435e26-4b25-483a-942d-477236992e84