Resolve a critical edge case in the XRL layer for the Microsoft Windows Server port of XORP 1.2.
The previous revision attempted to fix a regression in the Mac and
EtherMac classes by making the ether_aton() function (imported from
tcpdump) more liberal in the input which it accepts.

This commit however introduced the wrong behaviour. Both Mac and
EtherMac depend (on their definition) in the ability of ether_aton()
to parse Ethernet addresses rendered as ASCII strings in a manner
consistent across platforms, that is, the platform's ether_aton()
function must be bug-for-bug-compatible with FreeBSD's.

This in itself could be considered a design bug.

Because of this behaviour, whenever an ifmgr_mirror XRL client
registered its interface with the FEA, the updates which it received
were rejected as a non-critical XRL error (220 Internal error, corrupt
xrl) because the underlying Mac class used to parse the XRL mac address
atom was rejecting otherwise valid input.

Therefore, revert the previous commit, by introducing a FreeBSD-style
ether_aton() function which depends solely on the ISO C89 function
sscanf().

Approved by:	atanu (implicitly)
Reviewed by:	pavlin
Bugzilla ID:	http://www.xorp.org/bugzilla/show_bug.cgi?id=581


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8372 26435e26-4b25-483a-942d-477236992e84