Resolve a problem with launching binaries for the Microsoft Windows Server port of XORP 1.2.
On Windows, we cannot rely on the presence of a POSIX shell when
launching XORP system modules such as the RIB and FEA from the module
manager code in xorp_rtrmgr.exe.

In fact, the win_quote_args() function is intended to only be used
with non-UNIX style invocation, i.e. a direct call to CreateProcess().
Argument vectors passed to this function are munged in a manner which
allows them to be passed, as a single string argument for
CreateProcess(), to a Microsoft Visual C Runtime application, with
the correct quoting of backslashes and other Win32 metacharacters.

It should be considered whether RunShellCommand is needed at all,
and what it means to run something 'under the shell' on Win32.

Pavlin's patch is intended for a more general case but it is an
almost exact fix for the above issues under Win32.

Submitted by:	pavlin
Approved by:	atanu (implicitly)
Bugzilla ID:	http://www.xorp.org/bugzilla/show_bug.cgi?id=559
See also:	http://www.xorp.org/bugzilla/show_bug.cgi?id=505


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8373 26435e26-4b25-483a-942d-477236992e84