Make the Windows native event dispatch code more stable, by introducing two methods:  WinDispatcher::callback_bad_handle()  WinDispatcher::callback_bad_socket(XorpFd& fd)
Previously if we encountered a bad handle in
WinDispatcher::wait_and_dispatch(), we would blow
up with XLOG_FATAL(). Now we proactively dispatch
all callbacks registered on any bad handles in
the array, just like SelectorList does.

Sockets, however, are a special case. We rely on
WSAEnumNetworkEvents() to detect bad sockets, because
the handles registered and thus visible to
callback_bad_handle() are actually Event handles
used to dispatch events on each socket and not the
sockets themselves with which there is a 1:1
correspondence.

If we detect a bad socket, we do the same as for
a bad handle.

While we are here fix some nits in error messages
in the asserts for fatal edge cases around WFMO.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8392 26435e26-4b25-483a-942d-477236992e84