Fix for Windows native I/O event dispatcher.
Some code in the tree calls remove_ioevent_cb() with
a default argument of IOT_ANY for the type parameter.

With this change the socket event deregistration code
now knows how to deal with it whereas previously it
did not and would not remove any callbacks from a socket.

This does NOT fix the Windows BGP Hold Timer issue,
but it makes the condition non-fatal, thus allowing
a debugger to be attached.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8394 26435e26-4b25-483a-942d-477236992e84