Bug fix in checking the authentication sequence number. Previously, the check rejected a packet with sequence number same as the last accepted sequence number.
RFC 2082 says that the sequence number must be non-decreasing. The
exact text in Section 3.2.2. Message Reception is:

   If the neighbor has been heard
   from recently enough to have viable routes in the route table and the
   received sequence number is less than the last one received, the
   message likewise is discarded unprocessed.  When connectivity to the
   neighbor has been lost, the receiver SHOULD be ready to accept
   either:
   - a message with a sequence number of zero
   - a message with a higher sequence number than the last received
     sequence number.

The first sentence says "is less than" and covers the most common
case that is fixed with this commit.

Note that the last bullet says "a message with a higher sequence number".
This case actually applies only when the connectivity to a neighbor
is lost and assumes that we keep track of the last received sequence
number from such neighbor. In our implementation we don't keep state
about a neighbor if it times out hence we cannot perform such check.
Hence, for all practical purpose we will accept a first message
with any sequence number. After all, if a router is booted for a
first time, the sequence number in the first message from a neighbor
could have any value (in fact, the RFC doesn't seem to explicitly
cover this scenario).

Bugzilla URL:	http://www.xorp.org/bugzilla/show_bug.cgi?id=594
Submitted by:	Hasso Tepper


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8439 26435e26-4b25-483a-942d-477236992e84